requirements.txt
discord.py
aiohttp
websockets
chardet
sqlalchemy
asyncio